backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
yarg==0.1.9
